ipfs as class
